auto-detect bootstrapping in pg-translicator
When pg-translicator starts, if there is no data in the replica, it will assume that the system is bootstrapping and ask for changes from "0/0" (epoch / the beginning of time). However, if there is already data in the replica, it will pass an empty string as the LSN, which will just continue where it left off since last server restart.